feat: Add mock DCC services and async client
