+++ src/dcc_mcp_rpyc/testing/__init__.py
@@ -1,0 +1,10 @@
+"""Testing utilities for DCC-MCP-RPYC.
+
+This package provides testing utilities for DCC-MCP-RPYC, including mock services
+and test helpers for writing tests that interact with DCC applications.
+"""
+
+# Import local modules
+from dcc_mcp_rpyc.testing.mock_services import MockDCCService
+
+__all__ = ["MockDCCService"]
